# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for payment sandbox simulation.",
        "title": "Payment Sandbox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/merchant/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Merchant lists invoices with optional status and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoice"
                ],
                "summary": "List merchant invoices",
                "parameters": [
                    {
                        "enum": [
                            "PENDING",
                            "PAID",
                            "EXPIRED"
                        ],
                        "type": "string",
                        "description": "Invoice status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.Invoice"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.PaginationMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Merchant creates a new invoice. Requires Idempotency-Key header to safely retry on network failure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoice"
                ],
                "summary": "Create invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique key per logical request (UUID recommended). Replaying the same key returns the original response; reusing the key with a different body returns 409.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create invoice payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app_modules_invoice_handlers.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_invoice_handlers.InvoiceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "validation_error or idempotency_key_required",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "idempotency_key_conflict or idempotency_in_progress",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/merchant/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Merchant gets invoice detail by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invoice"
                ],
                "summary": "Get merchant invoice detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/app_modules_invoice_handlers.InvoiceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment-sandbox_app_shared_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app_modules_invoice_handlers.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "amount",
                "customer_email",
                "customer_name",
                "due_date"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 250000
                },
                "customer_email": {
                    "type": "string",
                    "example": "john.customer@example.com"
                },
                "customer_name": {
                    "type": "string",
                    "example": "John Customer"
                },
                "description": {
                    "type": "string",
                    "example": "Invoice for April subscription"
                },
                "due_date": {
                    "type": "string",
                    "example": "2026-05-01T10:00:00Z"
                }
            }
        },
        "app_modules_invoice_handlers.InvoiceResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "payment_link_token": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "payment-sandbox_app_modules_invoice_models_entity.Invoice": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "payment_link_token": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "payment-sandbox_app_modules_invoice_models_entity.InvoiceStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "PAID",
                "EXPIRED"
            ],
            "x-enum-varnames": [
                "InvoicePending",
                "InvoicePaid",
                "InvoiceExpired"
            ]
        },
...
        "payment-sandbox_app_shared_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/payment-sandbox_app_shared_response.ErrorPayload"
                },
                "meta": {}
            }
        },
        "payment-sandbox_app_shared_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "payment-sandbox_app_shared_response.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}